Update LMS repository URLs.
I plan to move the download site from the current solution to reduce cost (we're still running a Linux box to serve a few files...). In preparation of this move I slightly changed the update checks to make them less dependent of Logitech sponsored systems. The new links should work exactly as before.